requirements.txt
opencv-python~=4.7.0.72
vosk~=0.3.44
pyttsx3~=2.90
opencv-contrib-python~=4.7.0.72
requests~=2.28.0
numpy~=1.23.0


cython~=0.29.34

setuptools~=65.5.1
packaging~=23.0
pyparsing~=3.0.9






PyAudio~=0.2.13
dronebuddylib~=1.0.8
djitellopy~=2.5.0
mediapipe~=0.10.3
torch~=2.1.0
torchvision~=0.16.0
Pillow~=10.0.0
matplotlib~=3.7.2
pandas~=2.1.1
tqdm~=4.66.1
ultralytics~=8.0.199
google~=3.0.0
openai~=0.28.0
tiktoken~=0.5.1